Refactored signup form and finalized both auth page layouts and forms. Footer refactoring and validation next.
