+++ client/src/pages/auth/Login.js
@@ -1,12 +1,9 @@
 import React from 'react';
 import Auth from './Auth';
+import LoginForm from './components/LoginForm';
 
 const Login = () => {
-  return (
-    <Auth>
-      <h1>Login</h1>
-    </Auth>
-  );
+  return <Auth formComponent={<LoginForm />} title="Welcome Back!" />;
 };
 
 export default Login;
