+++ client/src/pages/auth/Signup.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import Auth from './Auth';
+import SignupForm from './components/SignupForm';
 
 const Signup = () => {
-  return <Auth />;
+  return <Auth formComponent={<SignupForm />} title="Welcome to KanBan!" />;
 };
 export default Signup;
